feat: add date format selection to settings
This will add a date format selection to the settings of the
application, it will also replace the default format converter of the
framework with a custom one.
This custom formatter will use a json configuration to get a pre
configured format. Those formats can be changed or extended inside the
resource directory of the application.

fix #8